Updated to latest sonar-scanner-cli image

+++ entrypoint.sh
@@ -13,6 +13,11 @@ if [[ -z "${SONAR_HOST_URL}" ]]; then
   exit 1
 fi
 
+if [[ -z "${INPUT_PROJECTBASEDIR}" ]]; then
+  echo "This GitHub Action requires the INPUT_PROJECTBASEDIR env variable."
+  exit 1
+fi
+
 if [[ -n "${SONAR_ROOT_CERT}" ]]; then
   echo "Adding custom root certificate to java certificate store"
   rm -f /tmp/tmpcert.pem
@@ -20,17 +25,22 @@ if [[ -n "${SONAR_ROOT_CERT}" ]]; then
   keytool -keystore /etc/ssl/certs/java/cacerts -storepass changeit -noprompt -trustcacerts -importcert -alias sonarqube -file /tmp/tmpcert.pem
 fi
 
-if [[ -f "${INPUT_PROJECTBASEDIR%/}pom.xml" ]]; then
+if [[ -f "${INPUT_PROJECTBASEDIR}/pom.xml" ]]; then
   echo "Maven project detected. You should run the goal 'org.sonarsource.scanner.maven:sonar' during build rather than using this GitHub Action."
   exit 1
 fi
 
-if [[ -f "${INPUT_PROJECTBASEDIR%/}build.gradle" ]]; then
+if [[ -f "${INPUT_PROJECTBASEDIR}/build.gradle" ]]; then
   echo "Gradle project detected. You should use the SonarQube plugin for Gradle during build rather than using this GitHub Action."
   exit 1
 fi
 
+if [[ -f "${INPUT_PROJECTBASEDIR}/pubspec.yaml" ]]; then
+  echo "Flutter project detected. Installing dependencies..."
+  cd $INPUT_PROJECTBASEDIR; flutter pub get
+fi
+
 unset JAVA_HOME
 
-sonar-scanner -Dsonar.projectBaseDir=${INPUT_PROJECTBASEDIR} ${INPUT_ARGS}
+cd $INPUT_PROJECTBASEDIR; sonar-scanner ${INPUT_ARGS}
